backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
python-multipart
sentence-transformers
faiss-cpu
numpy
